In Init.js de const updateBook aangepast. '&id=' toegevoegd en '+ bookId' toegevoegd. In update-book.js 'bookId' als 2e parameter meegegeven en de const bookId in update-book.php meegegeven.

+++ html/js/delete-book.js
@@ -5,7 +5,9 @@ document.addEventListener('DOMContentLoaded', () => {
 
 // Eventlistener aanmaken dat wanneer op de delete knop gedrukt wordt, de delete functie in de bookController geactiveerd wordt.
     form.addEventListener('submit', () => {
-        alert('Are you sure you want to delete this book?');
-        deleteBook(bookId);
+        confirmDelete = confirm('Are you sure you want to delete this book?');
+        if (confirmDelete === true) {
+            deleteBook(bookId);
+        }
     });
 });
